# catalog_swagger_jason.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/catalog/books":{"get":{"summary":"Get Books","operationId":"get_books_catalog_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Add Book","operationId":"add_book_catalog_books_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/books/{book_id}":{"get":{"summary":"Get Book","operationId":"get_book_catalog_books__book_id__get","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Book","operationId":"update_book_catalog_books__book_id__put","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Book","operationId":"delete_book_catalog_books__book_id__delete","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Book":{"title":"Book","type":"object","properties":{"identifier":{"title":"Identifier","type":"integer"},"title":{"title":"Title","type":"string"},"author":{"title":"Author","type":"string"},"year":{"title":"Year","type":"integer"},"quantity":{"title":"Quantity","type":"integer"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}